Midified: each user has own bucket

+++ Services/BucketService/BucketService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Book_Store.Data;
@@ -17,16 +18,21 @@ namespace Book_Store.Services.BucketService
     {
         private readonly IMapper mapper;
         public readonly DataContext context;
+        public IHttpContextAccessor httpContextAccessor;
 
-        public BucketService(IMapper mapper, DataContext context)
+        public BucketService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
+            this.httpContextAccessor = httpContextAccessor;
             this.context = context;
             this.mapper = mapper;
-        }         
+        }
+
+        private int GetUserId() => Convert.ToInt32(httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+
         public async Task<ServiceResponse<List<GetBucketItemsDto>>> GetBucketItems()
         {
             var serviceResponse = new ServiceResponse<List<GetBucketItemsDto>>();
-            var dbProducts = await context.Bucket.Include(c => c.Book).Include(c => c.Book.Author).Include(c => c.Book.Category).ToListAsync();
+            var dbProducts = await context.Bucket.Where(c => c.UserId == GetUserId()).Include(c => c.Book).Include(c => c.Book.Author).Include(c => c.Book.Category).ToListAsync();
             serviceResponse.Data = dbProducts.Select(c => mapper.Map<GetBucketItemsDto>(c)).ToList();
             return serviceResponse;
         }
@@ -37,7 +43,7 @@ namespace Book_Store.Services.BucketService
 
             try
             {
-                User user = await context.User.FirstOrDefaultAsync(c => c.Id == newBucketItem.UserId);
+                User user = await context.User.FirstOrDefaultAsync(c => c.Id == GetUserId());//newBucketItem.UserId);
                 if(user == null)
                 {
                     serviceResponse.Success = false;
@@ -53,10 +59,12 @@ namespace Book_Store.Services.BucketService
                 }
 
                 Bucket bucket = mapper.Map<Bucket>(newBucketItem);
+                bucket.UserId = GetUserId();
                 context.Bucket.Add(bucket);
                 await context.SaveChangesAsync();
 
-                serviceResponse.Data = await context.Bucket.Include(c => c.Book).Include(c => c.Book.Author).Include(c => c.Book.Category).Select(c => mapper.Map<GetBucketItemsDto>(c)).ToListAsync();
+                serviceResponse.Data = await context.Bucket.Where(c => c.UserId == GetUserId())
+                    .Include(c => c.Book).Include(c => c.Book.Author).Include(c => c.Book.Category).Select(c => mapper.Map<GetBucketItemsDto>(c)).ToListAsync();
             }
             catch(Exception ex)
             {
@@ -73,11 +81,21 @@ namespace Book_Store.Services.BucketService
 
             try
             {
-                var bucket = await context.Bucket.FirstAsync(c => c.Id == id);
-                context.Bucket.Remove(bucket);
-                await context.SaveChangesAsync();
+                var bucket = await context.Bucket.FirstAsync(c => c.Id == id && c.UserId == GetUserId());
+                if(bucket != null)
+                {
+                    context.Bucket.Remove(bucket);
+                    await context.SaveChangesAsync();
 
-                serviceResponse.Data = await context.Bucket.Include(c => c.Book.Category).Include(c => c.Book.Author).Select(c => mapper.Map<GetBucketItemsDto>(c)).ToListAsync();
+                    serviceResponse.Data = await context.Bucket.Where(c => c.UserId == GetUserId())
+                        .Include(c => c.Book.Category).Include(c => c.Book.Author).Select(c => mapper.Map<GetBucketItemsDto>(c)).ToListAsync();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Bucket item not found";
+                }
+
             }
             catch(Exception ex)
             {
